Handle case where token cache file contains invalid JSON
The case should actually be handled internally by the msal-node library
in TokenCache.deserialize() when deserialization of the JSON string
from the token cache file fails. Now the TokenCache is left
in an inconsistent state by the library.
We try to rollback to a consistent state, ie. resetting the token
cache, using library internals.